pyproject.toml
[project]
name = "esphome-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioesphomeapi>=41.7.0",
    "nabu-agent",
    "numpy>=2.3.3",
    "ovos-tts-plugin-matxa-multispeaker-cat>=0.0.1a5",
    "piper-tts>=1.3.0",
    "playsound3>=3.2.6",
    "pygame>=2.6.1",
    "sounddevice>=0.5.2",
    "zeroconf>=0.147.2",
]

[tool.uv.sources]
nabu-agent = { git = "https://github.com/und1n3/nabu-agent", rev = "whisper-test" }
